pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spot-downloader"
version = "0.1.0"
description = "Download Spotify playlists via YouTube Music in M4A format"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
keywords = ["spotify", "youtube", "music", "downloader", "m4a"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    # Spotify API
    "spotipy>=2.23.0",
    
    # YouTube Music API
    "ytmusicapi>=1.3.0",
    
    # YouTube download
    "yt-dlp>=2024.1.0",
    
    # Audio metadata
    "mutagen>=1.47.0",
    
    # Fuzzy string matching (same as spotDL)
    "rapidfuzz>=3.5.0",
    
    # CLI framework
    "click>=8.1.0",
    
    # Progress bars
    "tqdm>=4.66.0",
    
    # Configuration file parsing
    "pyyaml>=6.0",
    
    # Lyrics fetching (optional but included)
    "syncedlyrics>=0.7.0",
    
    # HTTP requests (for cover art, lyrics scraping)
    "requests>=2.31.0",
    
    # HTML parsing (for lyrics scraping)
    "beautifulsoup4>=4.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
]

[project.scripts]
spot = "spot_downloader.cli:main"

[project.urls]
Homepage = "https://github.com/Verryx-02/spot_downloader"
Documentation = "https://github.com/Verryx-02/spot_downloader#readme"
Repository = "https://github.com/Verryx-02/spot_downloader"
Issues = "https://github.com/Verryx-02/spot_downloader/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["spot_downloader*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["spot_downloader"]

[tool.black]
line-length = 100
target-version = ["py311"]
